+++ fgcm/fgcmSuperStarFlat.py
@@ -157,12 +157,12 @@ class FgcmSuperStarFlat(object):
                       (self.fgcmPars.expEpochIndex[obsExpIndex[goodObs2]],
                        self.fgcmPars.expLUTFilterIndex[obsExpIndex[goodObs2]],
                        obsCCDIndex[goodObs2]),
-                      1./EGrayErr2GO[mark])
+                      (1./EGrayErr2GO[mark]).astype(superStarWt.dtype))
             np.add.at(superStarOffset,
                       (self.fgcmPars.expEpochIndex[obsExpIndex[goodObs2]],
                        self.fgcmPars.expLUTFilterIndex[obsExpIndex[goodObs2]],
                        obsCCDIndex[goodObs2]),
-                      EGrayGO[mark]/EGrayErr2GO[mark])
+                      (EGrayGO[mark]/EGrayErr2GO[mark]).astype(superStarOffset.dtype))
             np.add.at(superStarNGoodStars,
                       (self.fgcmPars.expEpochIndex[obsExpIndex[goodObs2]],
                        self.fgcmPars.expLUTFilterIndex[obsExpIndex[goodObs2]],
